Promote ov066/Eyerok: fold 34 of 61 shards into src/actors/Eyerok.cpp
Text-only promotion. symbols.txt is untouched.

PARTIAL, and the queue's blocker was misfiled. `unmatched:6` reads as a
matching problem; it is a LINK problem. All six shards
(func_ov066_021166c8, _02116db0, _021175e8, _02117bf0, _02118188 and
_ZN6Eyerok13InitResourcesEv) byte-MATCH at 2004/b56 under tools/match.py
with --strict-relocs. They are excluded because they reference the
cross-overlay addresses 0x02112c08/c88/ca8/cc8/d48 under names that no
module's symbols.txt gives, so they carry no `complete` marker and dsd
leaves them out of the link entirely -- confirmed structurally: neither
src/func_ov066_021166c8.cpp nor src/_ZN6Eyerok13InitResourcesEv.cpp
appears in build/tu/_baseline/link/objects.txt.

Those six sit at ordinals 8, 16, 20, 22, 24 and 59, cutting the 61-function
run 0x02115ee0..0x0211a2e4 into pieces of 8/7/3/1/1/34/1. A licensed .text
claim cannot contain a hole, so this takes the largest contiguous linkable
side: ordinals 25..58, 0x021184c0..0x02119ce8. Dropped: 0..24, 59, 60.

Codegen. The file is written ROM-ASCENDING, which only works because of
`#pragma defer_codegen off` at the top. Three configurations were measured
over the identical 34-member source set:

  ROM-descending, no pragmas          33/34  (Behavior DIFF, 999 words)
  ROM-descending, file-global off     32/34  (021184e0 +4 words, 021194fc resized)
  ROM-ascending, defer off + bracket  34/34  MATCH

Turning deferral off does two things at once: it flips .text emission from
reverse-source to source order, and it makes the bracketed opt_common_subs /
opt_strength_reduction pair bind to _ZN6Eyerok8BehaviorEv alone instead of
leaking file-wide. Control: deleting the bracket outright takes Behavior
from 0x4b0 to 0x4ac. The pragmas are load-bearing, not decoration.

decl_common.h is deliberately NOT included. It types func_ov066_02119454 as
returning void; this TU defines that member and the ROM function returns a
value. Conforming to the header's spelling costs the match (measured both
ways with tools/match.py: int -> MATCH, void -> no match). Every symbol the
header would have supplied is declared explicitly instead.

The destructor stays out of line. The ROM puts D1 (0x02115ee0) below D0
(0x02115f84), which an in-class `virtual ~Eyerok() {}` would reproduce -- so
this class is not blocked on the destructor lever. But ordinals 0 and 1 are
outside this partial and still define it out of line, so include/Eyerok.h is
left unchanged. Consequence: this TU does not own the key function, emits no
vtable or typeinfo, and its compiler_only_output is 2 rows rather than the
13 a whole-run promotion would need.

Every one of the 34 members carries an `// @symbol` marker, C-derived ones
included, so tiers.py cannot charge an unmarked body to the member above it.
Adding the 29 missing markers is proved byte-neutral: the compiled object is
sha256 24e77c4f20c5848f21ac4316d5d02a09482f3fce1cfdc83dadfb5d0971deaee1,
18136 bytes, before and after.

Ratchet: 2 of this class's 5 CONVERTED baseline rows are inside the claim,
so they were banked, not skipped -- src/_ZN6Eyerok16CleanupResourcesEv.cpp
and src/_ZN6Eyerok16OnPendingDestroyEv.cpp become
src/actors/Eyerok.cpp#<member> rows. Count invariant at 2681, no unrelated
drift swept in.

Gates: verify 34/34 MATCH, objisolate clean, reloc-destinations clean;
linkcheck (pre-promotion) 34 LICENSED / 0 COLLIDES-GAP / 0 COLLIDES-SRC /
0 HOMELESS, emission order ROM-ascending as required, module fidelity
106/106, ROM sha256 identical to the stock control, dsd check symbols fails
only on the same 9 pre-existing errors as the baseline; layout_check clean;
eligible lists all 34; cpp_tu_compat --require-ready READY on all nine;
check_dead_references and check_tubuild_conflicts both green.
