+++ config/arm9/overlays/ov066/delinks.txt
@@ -98,141 +98,9 @@ src/func_ov066_02118168.c:
 src/func_ov066_02118188.c:
     .text start:0x02118188 end:0x021184c0
 
-src/func_ov066_021184c0.c:
+src/actors/Eyerok.cpp:
     complete
-    .text start:0x021184c0 end:0x021184e0
-
-src/func_ov066_021184e0.c:
-    complete
-    .text start:0x021184e0 end:0x021185e4
-
-src/func_ov066_021185e4.c:
-    complete
-    .text start:0x021185e4 end:0x02118604
-
-src/func_ov066_02118604.c:
-    complete
-    .text start:0x02118604 end:0x02118658
-
-src/func_ov066_02118658.c:
-    complete
-    .text start:0x02118658 end:0x02118678
-
-src/func_ov066_02118678.c:
-    complete
-    .text start:0x02118678 end:0x021187c8
-
-src/func_ov066_021187c8.cpp:
-    complete
-    .text start:0x021187c8 end:0x021188b0
-
-src/func_ov066_021188b0.c:
-    complete
-    .text start:0x021188b0 end:0x02118934
-
-src/func_ov066_02118934.c:
-    complete
-    .text start:0x02118934 end:0x02118954
-
-src/func_ov066_02118954.c:
-    complete
-    .text start:0x02118954 end:0x021189a0
-
-src/func_ov066_021189a0.c:
-    complete
-    .text start:0x021189a0 end:0x021189c0
-
-src/func_ov066_021189c0.c:
-    complete
-    .text start:0x021189c0 end:0x02118a30
-
-src/func_ov066_02118a30.c:
-    complete
-    .text start:0x02118a30 end:0x02118a50
-
-src/func_ov066_02118a50.c:
-    complete
-    .text start:0x02118a50 end:0x02118b08
-
-src/func_ov066_02118b08.c:
-    complete
-    .text start:0x02118b08 end:0x02118b28
-
-src/func_ov066_02118b28.c:
-    complete
-    .text start:0x02118b28 end:0x02118be0
-
-src/func_ov066_02118be0.c:
-    complete
-    .text start:0x02118be0 end:0x02118c00
-
-src/func_ov066_02118c00.c:
-    complete
-    .text start:0x02118c00 end:0x02118cb8
-
-src/func_ov066_02118cb8.c:
-    complete
-    .text start:0x02118cb8 end:0x02118cdc
-
-src/func_ov066_02118cdc.cpp:
-    complete
-    .text start:0x02118cdc end:0x02118de0
-
-src/func_ov066_02118de0.c:
-    complete
-    .text start:0x02118de0 end:0x02118e04
-
-src/func_ov066_02118e04.c:
-    complete
-    .text start:0x02118e04 end:0x0211901c
-
-src/func_ov066_0211901c.c:
-    complete
-    .text start:0x0211901c end:0x0211903c
-
-src/func_ov066_0211903c.c:
-    complete
-    .text start:0x0211903c end:0x02119348
-
-src/func_ov066_02119348.c:
-    complete
-    .text start:0x02119348 end:0x02119398
-
-src/func_ov066_02119398.cpp:
-    complete
-    .text start:0x02119398 end:0x0211944c
-
-src/func_ov066_0211944c.c:
-    complete
-    .text start:0x0211944c end:0x02119454
-
-src/func_ov066_02119454.cpp:
-    complete
-    .text start:0x02119454 end:0x021194a4
-
-src/func_ov066_021194a4.cpp:
-    complete
-    .text start:0x021194a4 end:0x021194fc
-
-src/func_ov066_021194fc.cpp:
-    complete
-    .text start:0x021194fc end:0x02119654
-
-src/_ZN6Eyerok16CleanupResourcesEv.cpp:
-    complete
-    .text start:0x02119654 end:0x021197a0
-
-src/_ZN6Eyerok16OnPendingDestroyEv.cpp:
-    complete
-    .text start:0x021197a0 end:0x021197a4
-
-src/_ZN6Eyerok6RenderEv.cpp:
-    complete
-    .text start:0x021197a4 end:0x02119838
-
-src/_ZN6Eyerok8BehaviorEv.cpp:
-    complete
-    .text start:0x02119838 end:0x02119ce8
+    .text start:0x021184c0 end:0x02119ce8
 
 src/_ZN6Eyerok13InitResourcesEv.cpp:
     .text start:0x02119ce8 end:0x0211a2dc
