+++ config/converted-backslide-exceptions.jsonl
@@ -346,3 +346,5 @@
 {"path": "src/_ZN12dScMgSound_cD1Ev.cpp", "reason": "ov006: promote dScMgSound_c into one 82-function C++ TU (src/actors/dScMgSound_c.cpp); the 82 one-function shards are absorbed, not lost -- linkcheck scratch-link-verified, 106/106 modules exact, ROM sha256 identical to stock"}
 {"path": "src/_ZN13dScMgTeresa_cD0Ev.cpp", "reason": "ov006/dScMgTeresa_c promotion: 48 one-function shards absorbed into src/actors/dScMgTeresa_c.cpp (ROM ordinals 0..47 of the 0x0211cbd0..0x021207a8 run; the other 33 stay enrolled behind a sourceless hole)"}
 {"path": "src/_ZN13dScMgTeresa_cD1Ev.cpp", "reason": "ov006/dScMgTeresa_c promotion: 48 one-function shards absorbed into src/actors/dScMgTeresa_c.cpp (ROM ordinals 0..47 of the 0x0211cbd0..0x021207a8 run; the other 33 stay enrolled behind a sourceless hole)"}
+{"path": "src/_ZN6Eyerok16CleanupResourcesEv.cpp", "reason": "ov066/Eyerok TU promotion: 34 of the 61-function run at 0x02115ee0..0x0211a2e4 fold into src/actors/Eyerok.cpp (.text 0x021184c0..0x02119ce8). Partial because six shards in the run reference cross-overlay addresses no symbols.txt names and so carry no complete marker; a licensed claim cannot contain a hole."}
+{"path": "src/_ZN6Eyerok16OnPendingDestroyEv.cpp", "reason": "ov066/Eyerok TU promotion: 34 of the 61-function run at 0x02115ee0..0x0211a2e4 fold into src/actors/Eyerok.cpp (.text 0x021184c0..0x02119ce8). Partial because six shards in the run reference cross-overlay addresses no symbols.txt names and so carry no complete marker; a licensed claim cannot contain a hole."}
