+++ config/converted-baseline.json
@@ -1365,9 +1365,7 @@
     "src/_ZN6Dorrie6RenderEv.cpp",
     "src/_ZN6DorrieD0Ev.cpp",
     "src/_ZN6DorrieD1Ev.cpp",
-    "src/_ZN6Eyerok16CleanupResourcesEv.cpp",
     "src/_ZN6Eyerok16OnAimedAtWithEggEv.cpp",
-    "src/_ZN6Eyerok16OnPendingDestroyEv.cpp",
     "src/_ZN6EyerokD0Ev.cpp",
     "src/_ZN6EyerokD1Ev.cpp",
     "src/_ZN6Fwoosh13OnTurnIntoEggER6Player.cpp",
@@ -2126,6 +2124,8 @@
     "src/actors/ActorDerived.cpp#_ZN7dBase_c5SpawnEjP7fBase_cii",
     "src/actors/ActorDerived.cpp#_ZN7dBase_cD0Ev",
     "src/actors/ActorDerived.cpp#_ZN7dBase_cD1Ev",
+    "src/actors/Eyerok.cpp#_ZN6Eyerok16CleanupResourcesEv",
+    "src/actors/Eyerok.cpp#_ZN6Eyerok16OnPendingDestroyEv",
     "src/actors/dBgActor_c.cpp#_ZN10dBgActor_c21UpdateModelPosAndRotYEv",
     "src/actors/dBgActor_c.cpp#_ZN10dBgActor_cC2Ev",
     "src/actors/dBgActor_c.cpp#_ZN10dBgActor_cD0Ev",
