--- src/_ZN6Eyerok16OnPendingDestroyEv.cpp
@@ -1,9 +1,0 @@
-//cpp
-// @symbol _ZN6Eyerok16OnPendingDestroyEv
-/* Eyerok::OnPendingDestroy -- vtable slot 12. The ROM body is empty: the
- * override exists only to occupy the slot. */
-#include "Eyerok.h"
-
-void Eyerok::OnPendingDestroy()
-{
-}
