--- src/func_ov066_02118a30.c
@@ -1,8 +1,0 @@
-int func_ov066_02118a30(char *p)
-{
-    *(int *)(p + 0x494) = 0;
-    *(int *)(p + 0x498) = 0;
-    *(short *)(p + 0x4d0) = 0;
-    *(int *)(p + 0x4a0) = 0;
-    return 1;
-}
