--- src/func_ov066_0211944c.c
@@ -1,4 +1,0 @@
-int func_ov066_0211944c(void)
-{
-    return 1;
-}
